% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioatles.R
\name{bioatles()}
\alias{bioatles()}
\alias{bioatles}
\title{Downloading species ocurrence data (presences) from Bioatles}
\usage{
bioatles(sp_dir = NULL, sp_list = NULL, out_name = "sp_records")
}
\arguments{
\item{sp_dir}{Directory of the species list}

\item{sp_list}{A csv file name or a vector containing species to be downloaded}

\item{out_name}{Name to the output data set (csv file)}
}
\value{
An csv file with the occurrences in Lat/Long Geographic Coordinates System WGS84.
}
\description{
Download species (presences) data from the Bioatles http://bioatles.caib.es. As data in Bioatles is projected in European Datum 1950 (31N), the function also trensform it in Lat/Long Geographic Coordinates System WGS84. bioatles() is based on several functions included in the packages "rvest" (Wickham, 2016) and "xml2" (Wickham et al., 2017). Finally, it saves the data in a csv file.
}
\details{
The aim of this script is to define the function bioatles(), which is used to download species (presences) data from the Bioatles http://bioatles.caib.es. As data in Bioatles is projected in European Datum 1950 (31N), the function also trensform it in Lat/Long Geographic Coordinates System WGS84. bioatles() is based on several functions included in the packages "rvest" (Wickham, 2016) and "xml2" (Wickham et al., 2017). Finally, it saves the data in a csv file.
}
\author{
Xavier Rotllan-Puig
}
