% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gbif_preparation.R
\name{PrepBIF()}
\alias{PrepBIF()}
\alias{Prep_BIF}
\title{GetBIF}
\usage{
Prep_BIF(
  taxon_dir = NULL,
  taxons = NULL,
  cols2keep = c("species", "decimalLatitude", "decimalLongitude"),
  rm_dupl = NULL
)
}
\arguments{
\item{cols2keep}{Column names to keep in the final data set. Default, cols2keep = c("species", "decimalLatitude", "decimalLongitude"). cols2keep = "all", keeps all columns}

\item{rm_dupl}{If TRUE (default), duplicate occurrences (same sp, same coordinates) are removed from the final data set}
}
\value{
A data frame
}
\description{
Retrieving the data set(s) downloaded with GetBIG() and creating a simpler data set (csv) with only those fields decided by the user. This function is used by GetBIF()
}
\details{
Preparing the species ocurrences data set from GBIF to be saved as a simpler csv
}
\examples{
\donttest{
Prep_BIF(taxon_dir = NULL,
         taxons = NULL,
         cols2keep = c("species", "decimalLatitude", "decimalLongitude"),
         rm_dupl = NULL
         )
}

}
\author{
Xavier Rotllan-Puig
}
